wh65lp_rs485_mqtt_bridge/requirements.txt
paho-mqtt>=2.0.1,<3.0.0
